+++ RaspberryPiAPI/src/main/java/sloshbot/raspberrypi_api/repositories/OrderHistoryRepository.java
@@ -3,5 +3,5 @@ package sloshbot.raspberrypi_api.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sloshbot.raspberrypi_api.models.DAOs.OrderHistory;
 
-public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Long> {
+public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Integer> {
 }
